Added reference designator check
Script now scans all reference designators and makes note of any ones which do not follow the standard format. Intended to quickly highlight if the user has mistakenly used the "Reference" field to write a comment, or overwritten it with a nonstandard part name (which is what the "name" field is for).